switch atsign impl
