+++ at_client/lib/src/listener/at_sign_change_listener.dart
@@ -1,5 +1,5 @@
-import 'package:at_client/at_client.dart';
+import 'package:at_client/src/listener/switch_at_sign_event.dart';
 
 abstract class AtSignChangeListener {
-  void listenToAtSignChange(AtClient atClient);
+  void listenToAtSignChange(SwitchAtSignEvent switchAtSignEvent);
 }
