+++ at_client/lib/src/listener/switch_at_sign_event.dart
@@ -1,0 +1,7 @@
+import 'package:at_client/at_client.dart';
+
+class SwitchAtSignEvent {
+  AtClient? previousAtClient;
+  late AtClient newAtClient;
+  SwitchAtSignEvent(this.previousAtClient, this.newAtClient);
+}
