+++ at_client/lib/src/manager/at_client_manager.dart
@@ -1,9 +1,14 @@
 import 'package:at_client/at_client.dart';
 import 'package:at_client/src/listener/at_sign_change_listener.dart';
+import 'package:at_client/src/listener/switch_at_sign_event.dart';
+import 'package:at_client/src/service/notification_service_impl.dart';
+import 'package:at_client/src/service/sync_service_impl.dart';
 
 class AtClientManager {
   var _atSign;
-  var currenAtClient;
+  late var atClient;
+  late var syncService;
+  late var notificationService;
   final _changeListeners = <AtSignChangeListener>[];
 
   static final AtClientManager _singleton = AtClientManager._internal();
@@ -17,17 +22,21 @@ class AtClientManager {
   void setCurrentAtSign(
       String atSign, String namespace, AtClientPreference preference) {
     _atSign = _atSign;
-    currenAtClient = AtClientImpl(_atSign, namespace, preference);
-    _notifyListeners(currenAtClient);
+    final previousAtClient = atClient;
+    atClient = AtClientImpl.create(_atSign, namespace, preference);
+    final switchAtSignEvent = SwitchAtSignEvent(previousAtClient, atClient);
+    syncService = SyncServiceImpl.create(atClient);
+    notificationService = NotificationServiceImpl.create(atClient);
+    _notifyListeners(switchAtSignEvent);
   }
 
   void listenToAtSignChange(AtSignChangeListener listener) {
     _changeListeners.add(listener);
   }
 
-  void _notifyListeners(AtClient atClient) {
+  void _notifyListeners(SwitchAtSignEvent switchAtSignEvent) {
     _changeListeners.forEach((listener) {
-      listener.listenToAtSignChange(atClient);
+      listener.listenToAtSignChange(switchAtSignEvent);
     });
   }
 }
